Problem 4: Optimized algorithm; removed Log10()
In this particular case, we know that we should find a palindrome formed
by two three-digits number, so we're looking for x as:

100×100 ≤ x ≤ 999×999
10000 ≤ x ≤ 998001

So we're looking at a five or six digits number.
Given this, we can afford a single sub-nanosecond comparison instead of
a base-10 logarithm executed in nearly 20ns.

The remaining of the algorithm was improved through an empirical method,
benchmarking each modification in order to find the best way of caching
values to registers, looping, testing conditions, …

A parallelized algorithm was too expensive for synchronization due to
mutexes locks and unlocks (reaching hundreds of microseconds)